Baseline of creating final as a component for 5b
A new component is added that takes x, y, and z as inputs then outputs the final value of x,y, and z respectively at every node. The goal is for this to replace the independent variable that we were originally using. Other wise there are three other options: omit the constraint, modify the independent thing Shugo said to do, see if we can give the full Jacobian matrix and derive a more involved partial and inserting the entire Jacobian which I'm not even sure is possible. Low key just replace with the analytical gradient if this doesn't work